# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kang Raling API Web Service",
...
        "contact": {
            "name": "Developer Website Kang Raling",
            "email": "danu3anugrah@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server Lokal Development"
        },
        {
            "url": "https://kangraling.dlhgarut.id",
            "description": "Server Utama (Production Dinas Lingkungan Hidup)"
        }
    ],
    "paths": {
...
        "/desas": {
            "get": {
                "tags": [
                    "Publik"
                ],
                "summary": "Ambil semua desa binaan",
                "description": "Mengembalikan daftar semua desa binaan DLH Kabupaten Garut beserta profil TPS-nya. Tidak membutuhkan token — endpoint publik.",
                "operationId": "d34174560b8cc0ea54cb82cc5883776e",
                "responses": {
                    "200": {
                        "description": "Daftar desa binaan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama_desa": {
                                                        "type": "string",
                                                        "example": "Desa Sukamaju"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "desa-sukamaju"
                                                    },
                                                    "alamat": {
                                                        "type": "string",
                                                        "example": "Jl. Raya Garut No. 1",
                                                        "nullable": true
                                                    },
                                                    "profil_tps": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Desa Binaan"
                ],
                "summary": "Tambah desa binaan baru",
                "description": "Menambahkan desa binaan baru ke sistem. Membutuhkan Auth Token dengan permission `kelola.desa-binaan`.",
                "operationId": "0599ab1036ef249fabeabc7178d25791",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_desa"
                                ],
                                "properties": {
                                    "nama_desa": {
                                        "type": "string",
                                        "example": "Desa Sukamaju"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Raya Garut No. 1, Kecamatan Tarogong",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Desa berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Desa berhasil ditambahkan."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nama_desa": {
                                                    "type": "string",
                                                    "example": "Desa Sukamaju"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "desa-sukamaju"
                                                },
                                                "alamat": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Tidak punya permission kelola.desa-binaan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/desas/{slug}": {
            "get": {
                "tags": [
                    "Publik"
                ],
                "summary": "Ambil detail desa binaan berdasarkan slug",
                "description": "Mengembalikan detail lengkap satu desa binaan beserta profil TPS-nya. Tidak membutuhkan token.",
                "operationId": "fda262d5c108772768302f245a1f5788",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug nama desa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "desa-sukamaju"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail desa berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nama_desa": {
                                                    "type": "string",
                                                    "example": "Desa Sukamaju"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "desa-sukamaju"
                                                },
                                                "alamat": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "profil_tps": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Desa tidak ditemukan"
                    }
                }
            }
        },
        "/desas/{id}": {
            "put": {
                "tags": [
                    "Desa Binaan"
                ],
                "summary": "Update data desa binaan",
                "description": "Mengupdate nama dan alamat desa binaan. Slug akan otomatis di-generate ulang dari nama desa yang baru.",
                "operationId": "3e689db8a1d5bc85eb4ea3bf84200b61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID desa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_desa"
                                ],
                                "properties": {
                                    "nama_desa": {
                                        "type": "string",
                                        "example": "Desa Sukamaju Baru"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Raya Garut No. 2",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Desa berhasil diperbarui"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Tidak punya permission kelola.desa-binaan"
                    },
                    "404": {
                        "description": "Desa tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Desa Binaan"
                ],
                "summary": "Hapus desa binaan",
                "description": "Menghapus desa binaan secara permanen. Khusus Administrator dengan permission `kelola.desa-binaan`.",
                "operationId": "507c5c953ef5c11f2cab4c9a13bf64e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Desa berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Desa berhasil dihapus."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Tidak punya permission kelola.desa-binaan"
                    },
                    "404": {
                        "description": "Desa tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token dengan format: **Bearer {token}**",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}